+++ package.json
@@ -1,12 +1,15 @@
 {
   "name": "ts-type-diagnostic",
-  "version": "0.0.18",
+  "version": "0.0.21",
   "main": "./bin/cli.js",
   "bin": {
     "tstd": "./bin/cli.js"
   },
   "scripts": {
-    "build": "tsc -p .",
+    "build": "npm run clean && tsc -p .",
+    "clean": "rm -rf bin",
+    "test": "npm run jest --",
+    "jest": "jest",
     "tstd": "ts-node ./cli.ts"
   },
   "repository": {
@@ -35,6 +38,9 @@
     "ts-node": "10.9.1",
     "typescript": "4.9.5",
     "yargs": "^17.4.1"
+  },
+  "devDependencies": {
+    "jest": "28.1.3"
   },
   "prettier": {
     "printWidth": 120,
